Started working on updates to the README's with new file structure and updates to the python script usages
Reformatted the example directories so there are "prepackaged" and "empty" directories -- one for the user to run simulations in, and one for them to compare against

Found a bug in coreFuncs.py and adjusted it -- if a "/" is provided with os.join, it struggles to join the path correctly

Still TO-DO:  coreFuncs.py assumes that the wave data is stored in an "interfaceHeight1" directory, but we currently have all the data being saved to a "waveProbes" directory, so that will need to be adjusted so that things work correctly. Updates to the README are still a WIP. Need to remove all the README duplication within the mesh convergence study folders.